Add migration-tool for aufs to overlay2
Implements storage-driver migration without the need to pull layer data again.
To achieve this we translate the on-disk AUFS file tree to the format used by
overlayFS. This includes containers that exist at the time of the migration.

The major difference between the two implementations of stacking FS is
that AUFS makes use of "marker files" to track deleted files in the
upper layer, while overlay employs file attributes and character files.
The other part of the work is reorganizing the data on disk to adhere to
the structure the overlay2 driver expects.

To make these changes effective after running a migration the engine has to
be reconfigured to use the overlay2 storage-driver and containers have to be
restarted.

The code in cmd/a2o-migrate can be used stand-alone. For that see the
makefile "a2o-migrate" target.

Change-type: patch
Signed-off-by: Robert Günzler <robertg@balena.io>